+++ CesiumRasterOverlays/include/CesiumRasterOverlays/ActivatedRasterOverlay.h
@@ -262,8 +262,8 @@ public:
   /**
    * @brief Ticks the underlying raster overlay. This should be called
    * frequently, such as once per frame, so that raster overlays can perform
-   * necessary updates. `Tileset` will call this method on all attached raster
-   * overlays automatically.
+   * necessary updates. \ref Cesium3DTilesSelection::Tileset will call this
+   * method on all attached raster overlays automatically.
    */
   void tick();